Create chat subfolder to better manage chat related modules
Needed for future commits.
