+++ src/meson.build
@@ -27,10 +27,11 @@ sources = files(
   'main.rs',
   'utils.rs',
   'window.rs',
-  'session/chat.rs',
   'session/chat_list.rs',
   'session/mod.rs',
   'session/session.rs',
+  'session/chat/chat.rs',
+  'session/chat/mod.rs',
   'session/content/chat_history.rs',
   'session/content/content.rs',
   'session/content/mod.rs',
