+++ src/session/chat/mod.rs
@@ -1,0 +1,3 @@
+mod chat;
+
+pub use self::chat::Chat;
